Add useUserPresence hook
