+++ frontend/src/components/headers/DocumentHeader.tsx
@@ -15,70 +15,26 @@ import {
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useNavigate } from "react-router-dom";
-import { useList } from "react-use";
-import { ActorID } from "yorkie-js-sdk";
+import { useUserPresence } from "../../hooks/useUserPresence";
 import { EditorModeType, selectEditor, setMode } from "../../store/editorSlice";
 import { selectWorkspace } from "../../store/workspaceSlice";
-import { YorkieCodeMirrorPresenceType } from "../../utils/yorkie/yorkieSync";
 import DownloadMenu from "../common/DownloadMenu";
 import ShareButton from "../common/ShareButton";
 import ThemeButton from "../common/ThemeButton";
 import UserPresenceList from "./UserPresenceList";
-
-export type Presence = {
-	clientID: ActorID;
-	presence: YorkieCodeMirrorPresenceType;
-};
 
 function DocumentHeader() {
 	const dispatch = useDispatch();
 	const navigate = useNavigate();
 	const editorState = useSelector(selectEditor);
 	const workspaceState = useSelector(selectWorkspace);
-	const [
-		presenceList,
-		{
-			set: setPresenceList,
-			push: pushToPresenceList,
-			removeAt: removePresenceAt,
-			clear: clearPresenceList,
-			filter: filterPresenceList,
-		},
-	] = useList<Presence>([]);
+	const { presenceList } = useUserPresence(editorState.doc);
 
 	useEffect(() => {
 		if (editorState.shareRole === "READ") {
 			dispatch(setMode("read"));
 		}
 	}, [dispatch, editorState.shareRole]);
-
-	useEffect(() => {
-		if (!editorState.doc) return;
-
-		setPresenceList(editorState.doc.getPresences());
-
-		const unsubscribe = editorState.doc.subscribe("others", (event) => {
-			if (event.type === "watched") {
-				setPresenceList(editorState.doc?.getPresences?.() ?? []);
-			}
-
-			if (event.type === "unwatched") {
-				filterPresenceList((presence) => presence.clientID !== event.value.clientID);
-			}
-		});
-
-		return () => {
-			unsubscribe();
-			clearPresenceList();
-		};
-	}, [
-		editorState.doc,
-		clearPresenceList,
-		pushToPresenceList,
-		removePresenceAt,
-		setPresenceList,
-		filterPresenceList,
-	]);
 
 	const handleChangeMode = (newMode: EditorModeType) => {
 		if (!newMode) return;
